+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2.6.4
 
+ - Load player skins asynchronously, and various fixes to the camera spell.
+
 ## 2.6.3
 
  - Make the undo system use batched block updates